LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/GUM-api.https-expected.txt

Description

This test checks for the presence of the navigator.mediaDevices.getUserMedia method.


PASS mediaDevices.getUserMedia() is present on navigator
